Cleaning up the prompt files
