+++ packages/cli/commands/create/api-sample.js
@@ -1,5 +1,9 @@
-const { createApiSamplePrompt } = require('../../lib/createApiSamplePrompt');
-const { folderOverwritePrompt } = require('../../lib/prompts');
+const {
+  createApiSamplePrompt,
+} = require('../../lib/prompts/createApiSamplePrompt');
+const {
+  folderOverwritePrompt,
+} = require('../../lib/prompts/folderOverwritePrompt');
 const { logger } = require('@hubspot/cli-lib/logger');
 const path = require('path');
 const fs = require('fs-extra');
