+++ packages/cli/lib/prompts/folderOverwritePrompt.js
@@ -1,0 +1,14 @@
+const { promptUser } = require('./promptUtils');
+
+const folderOverwritePrompt = folderName => {
+  return promptUser({
+    type: 'confirm',
+    name: 'overwrite',
+    message: `The folder with name '${folderName}' already exists. Overwrite?`,
+    default: false,
+  });
+};
+
+module.exports = {
+  folderOverwritePrompt,
+};
