+++ packages/cli/lib/prompts/promptUtils.js
@@ -1,0 +1,10 @@
+const inquirer = require('inquirer');
+
+const promptUser = async promptConfig => {
+  const prompt = inquirer.createPromptModule();
+  return prompt(promptConfig);
+};
+
+module.exports = {
+  promptUser,
+};
